+++ app/src/content_scripts/webApps/youtube/components/TogglePlaybackRateButton.tsx
